Migrate Query.clientStatsByTargets
