+++ packages/services/api/src/modules/operations/resolvers.ts
@@ -6,40 +6,6 @@ import { OperationsManager } from './providers/operations-manager';
 
 export const resolvers: OperationsModule.Resolvers = {
   Query: {
-    async clientStatsByTargets(_, { selector }, { injector }) {
-      const translator = injector.get(IdTranslator);
-      const [organization, project] = await Promise.all([
-        translator.translateOrganizationId(selector),
-        translator.translateProjectId(selector),
-      ]);
-
-      const targets = selector.targetIds;
-      const period = parseDateRangeInput(selector.period);
-
-      const [rows, total] = await Promise.all([
-        injector.get(OperationsManager).readUniqueClientNames({
-          target: targets,
-          project,
-          organization,
-          period,
-        }),
-        injector.get(OperationsManager).countRequests({
-          organization,
-          project,
-          target: targets,
-          period,
-        }),
-      ]);
-
-      return rows.map(row => {
-        return {
-          name: row.name,
-          count: row.count,
-          percentage: total === 0 ? 0 : (row.count / total) * 100,
-          versions: [], // TODO: include versions at some point
-        };
-      });
-    },
     async monthlyUsage(_, { selector }, { injector }) {
       const translator = injector.get(IdTranslator);
       const organization = await translator.translateOrganizationId(selector);
